requirements.txt
torch
openai-whisper==20240930
tqdm>=4.65.0
requests>=2.31.0
pyinstaller==6.3.0
ffmpeg-python
soundfile
numpy